Storing flyer data & accessing it in the main screen
I stored all of the text information inputted into the flyer creation activity into a JSON object, which stores titles and strings. Later we can use the JSON obj for listing each flyer created and for use in the flyer details screen.

I also made it so that the picture taken by the camera would carry over to the main activity screen, via a .png file. (Which we're probably going to want to change for the final app, since it gets rewritten every time a picture is taken.) This was mostly just to test/figure out how to use information inputted in one activity in another.