CardView/MainActivity layout displaying flyers
Now when you upload a flyer it will save all the information to the firebase storage & database. The MainActivity layout now shows each uploaded flyer as a cardview with an image and a title/description. I wasn't sure how user login & authentication factored in so I commented out the code for now.

I found a great tutorial that explains everything:
https://www.youtube.com/watch?v=lPfQN-Sfnjw&t=338s

Things to do:
-have a detailed view when you select a flyer
-there's an empty card at the bottom, I'm not sure how to fix that.
-add a title textthing in the upload activity
-user stuff